+++ sources/sdk/rust/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "drasi-source-sdk"
-version = "0.1.7"
+version = "0.1.11"
 edition = "2021"
 license = "Apache-2.0"
 description = "Source SDK for Drasi"
