+++ sources/sdk/rust/example/proxy/Cargo.toml
@@ -5,5 +5,5 @@ edition = "2021"
 
 [dependencies]
 tokio = {version = "1.40.0", features = ["full"]}
-drasi-source-sdk = "0.1.2"
+drasi-source-sdk = "0.1.8"
 serde_json = "1.0.133"
