+++ sources/sdk/rust/example/reactivator/Cargo.toml
@@ -5,6 +5,6 @@ edition = "2021"
 
 [dependencies]
 tokio = {version = "1.40.0", features = ["full"]}
-drasi-source-sdk = "0.1.7"
+drasi-source-sdk = "0.1.10"
 serde_json = "1.0.133"
 log = "0.4.22"
